dom/xml/test/test_viewport.xhtml

<?xml version="1.0"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=436083
-->
<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
  <meta name="viewport" content="height=398, width=4224, minimum-scale=0.1,
  initial-scale=2.3, maximum-scale=45.2, user-scalable=no" />
  <title>Test for Viewport META Tag Parsing</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank"
href="https://bugzilla.mozilla.org/show_bug.cgi?id=436083">Mozilla Bug 436083</a>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

/** Test for Viewport META Tag **/

SimpleTest.waitForExplicitFinish();

function testViewport() {

  /* Grab Viewport Metadata from the document header. */
  var windowUtils = SpecialPowers.getDOMWindowUtils(window);
  var vpWidth =
    parseInt(windowUtils.getDocumentMetadata("viewport-width"));
  var vpHeight =
    parseInt(windowUtils.getDocumentMetadata("viewport-height"));
  var vpInitialScale = 
    parseFloat(windowUtils.getDocumentMetadata("viewport-initial-scale"));
  var vpMaxScale = 
    parseFloat(windowUtils.getDocumentMetadata("viewport-maximum-scale"));
  var vpMinScale = 
    parseFloat(windowUtils.getDocumentMetadata("viewport-minimum-scale"));
  var vpUserScalable = 
    windowUtils.getDocumentMetadata("viewport-user-scalable");

  is(vpWidth, 4224, "Should get proper width");
  is(vpHeight, 398, "Should get proper height");
  is(vpInitialScale, 2.3, "Should get proper initial scale");
  is(vpMaxScale, 45.2, "Should get proper max scale");
  is(vpMinScale, 0.1, "Should get proper min scale");
  is(vpUserScalable, "no", "Should get proper user scalable parameter");
}

addLoadEvent(testViewport);
addLoadEvent(SimpleTest.finish);
]]>

</script>
</pre>
</body>
</html>
